refactor: 컴포넌트의 인덱싱을 위해 index.tsx, 컴포넌트.tsx로 분리했던 파일 구조 컨벤션을 index.tsx 하나로 관리하도록 변경
