--- components/Admin/Nav/Nav.tsx
@@ -1,15 +1,0 @@
-import { ADMIN } from '@/constants';
-import * as S from './Nav.styled';
-
-const Nav = () => {
-  return (
-    <S.HeaderContainer href={ADMIN.HOME}>
-      <S.Header>
-        <S.LogoContainer />
-        Leets
-      </S.Header>
-    </S.HeaderContainer>
-  );
-};
-
-export default Nav;
