+++ components/Admin/Nav/index.tsx
@@ -1,3 +1,15 @@
-import Nav from './Nav';
+import { ADMIN } from '@/constants';
+import * as S from './Nav.styled';
+
+const Nav = () => {
+  return (
+    <S.HeaderContainer href={ADMIN.HOME}>
+      <S.Header>
+        <S.LogoContainer />
+        Leets
+      </S.Header>
+    </S.HeaderContainer>
+  );
+};
 
 export default Nav;
