--- components/BackgroundImage/BackgroundImage.tsx
@@ -1,99 +1,0 @@
-import { memo } from 'react';
-
-const BackgroundImage = () => {
-  return (
-    <svg viewBox="0 0 1440 912" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <g clipPath="url(#clip0_100_39)">
-        <g filter="url(#filter0_f_100_39)">
-          <path
-            d="M720 195.104C647.496 195.104 506.885 312.731 420.426 161.173C362.225 59.1502 293.5 151 271.032 173.382C248.564 195.763 184 225.448 144 200C104 174.552 87.6439 136.157 73.2876 102.055C51.6613 50.6832 36.0483 13.5955 -46 0V410H1486V0C1403.95 13.5955 1386 72 1370.5 92C1355 112 1344 131.827 1310 153.5C1276 175.173 1205.93 173.024 1179 153.5C1136 122.327 1056.93 127.827 1024.5 153.5C887.7 261.81 792.504 195.104 720 195.104Z"
-            fill="url(#paint0_linear_100_39)"
-          />
-        </g>
-        <g filter="url(#filter1_f_100_39)">
-          <path
-            d="M468 218.601C233 338.765 79.8333 252.434 -15 218.601V381.601H1456V174.601C1429 222.101 1140 323.601 938.5 206.601C737 89.6006 666.5 117.101 468 218.601Z"
-            fill="url(#paint1_linear_100_39)"
-          />
-        </g>
-        <rect y="328" width="1440" height="584" fill="url(#paint2_linear_100_39)" />
-        <g filter="url(#filter2_f_100_39)">
-          <path
-            d="M0 226.55C0 226.55 120.5 239.461 284 239.461C447.5 239.461 475 227 724 223.397C973 219.794 1010.5 241.825 1150 244.977C1289.5 248.13 1440 226.55 1440 226.55V401H0V226.55Z"
-            fill="url(#paint3_linear_100_39)"
-          />
-        </g>
-      </g>
-      <defs>
-        <filter
-          id="filter0_f_100_39"
-          x="-110"
-          y="-64"
-          width="1660"
-          height="538"
-          filterUnits="userSpaceOnUse"
-          colorInterpolationFilters="sRGB">
-          <feFlood floodOpacity="0" result="BackgroundImageFix" />
-          <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape" />
-          <feGaussianBlur stdDeviation="32" result="effect1_foregroundBlur_100_39" />
-        </filter>
-        <filter
-          id="filter1_f_100_39"
-          x="-39"
-          y="106"
-          width="1519"
-          height="299.601"
-          filterUnits="userSpaceOnUse"
-          colorInterpolationFilters="sRGB">
-          <feFlood floodOpacity="0" result="BackgroundImageFix" />
-          <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape" />
-          <feGaussianBlur stdDeviation="12" result="effect1_foregroundBlur_100_39" />
-        </filter>
-        <filter
-          id="filter2_f_100_39"
-          x="-30"
-          y="193"
-          width="1500"
-          height="238"
-          filterUnits="userSpaceOnUse"
-          colorInterpolationFilters="sRGB">
-          <feFlood floodOpacity="0" result="BackgroundImageFix" />
-          <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape" />
-          <feGaussianBlur stdDeviation="15" result="effect1_foregroundBlur_100_39" />
-        </filter>
-        <linearGradient
-          id="paint0_linear_100_39"
-          x1="337"
-          y1="105.072"
-          x2="339.496"
-          y2="358.047"
-          gradientUnits="userSpaceOnUse">
-          <stop stopColor="#3685FC" stopOpacity="0" />
-          <stop offset="0.473958" stopColor="#3685FC" stopOpacity="0.588542" />
-          <stop offset="1" stopColor="#3685FC" />
-        </linearGradient>
-        <linearGradient id="paint1_linear_100_39" x1="350" y1="170" x2="1083.5" y2="170" gradientUnits="userSpaceOnUse">
-          <stop stopColor="#3685FC" stopOpacity="0" />
-          <stop offset="0.177083" stopColor="#3685FC" stopOpacity="0.738636" />
-          <stop offset="0.458333" stopColor="#3685FC" />
-          <stop offset="0.53125" stopColor="#3685FC" />
-          <stop offset="0.817708" stopColor="#3685FC" stopOpacity="0.711111" />
-          <stop offset="1" stopColor="#3685FC" stopOpacity="0" />
-        </linearGradient>
-        <linearGradient id="paint2_linear_100_39" x1="720" y1="328" x2="720" y2="912" gradientUnits="userSpaceOnUse">
-          <stop stopColor="#3685FC" stopOpacity="0" />
-          <stop offset="0.0885417" stopColor="#3685FC" />
-        </linearGradient>
-        <linearGradient id="paint3_linear_100_39" x1="720" y1="246.91" x2="720" y2="401" gradientUnits="userSpaceOnUse">
-          <stop stopColor="#3685FC" stopOpacity="0" />
-          <stop offset="1" stopColor="#3685FC" />
-        </linearGradient>
-        <clipPath id="clip0_100_39">
-          <rect width="1440" height="912" fill="white" />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-};
-
-export default memo(BackgroundImage);
