+++ components/Footer/index.tsx
@@ -1,3 +1,13 @@
-import Footer from './Footer';
+import { memo } from 'react';
+import * as S from './Footer.styled';
 
-export default Footer;
+const Footer = () => (
+  <S.FooterContainer>
+    <S.ContentContainer>
+      <S.HrStyle />
+      <S.CopyrightText>Copyright 2023. Leets all rights reserved.</S.CopyrightText>
+    </S.ContentContainer>
+  </S.FooterContainer>
+);
+
+export default memo(Footer);
