--- components/Promotions/Timeline/TimelineImage/TimelineImage.tsx
@@ -1,27 +1,0 @@
-import * as S from './TimelineImage.styled';
-
-const TimelineImage = () => (
-  <S.TimelineImageContainer>
-    <svg viewBox="0 0 36 978" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <line x1="18" y1="8.74228e-08" x2="18" y2="960" stroke="url(#paint0_linear_343_528)" strokeWidth="4" />
-      <circle cx="18" cy="196" r="18" fill="#3685FC" fillOpacity="0.2" />
-      <circle cx="18" cy="196" r="6" fill="#3685FC" />
-      <circle cx="18" cy="387" r="18" fill="#3685FC" fillOpacity="0.2" />
-      <circle cx="18" cy="387" r="6" fill="#3685FC" />
-      <circle cx="18" cy="960" r="18" fill="#3685FC" fillOpacity="0.2" />
-      <circle cx="18" cy="960" r="6" fill="#3685FC" />
-      <circle cx="18" cy="769" r="18" fill="#3685FC" fillOpacity="0.2" />
-      <circle cx="18" cy="769" r="6" fill="#3685FC" />
-      <circle cx="18" cy="578" r="18" fill="#3685FC" fillOpacity="0.2" />
-      <circle cx="18" cy="578" r="6" fill="#3685FC" />
-      <defs>
-        <linearGradient id="paint0_linear_343_528" x1="16" y1="960" x2="15.9999" y2="0" gradientUnits="userSpaceOnUse">
-          <stop stopColor="#3685FC" />
-          <stop offset="1" stopColor="#3685FC" stopOpacity="0" />
-        </linearGradient>
-      </defs>
-    </svg>
-  </S.TimelineImageContainer>
-);
-
-export default TimelineImage;
